add jest and utilize packr
