+++ .eslintrc.json
@@ -2,12 +2,14 @@
   "env": {
     "node": true,
     "commonjs": true,
-    "es2021": true
+    "es2021": true,
+    "jest/globals": true
   },
   "extends": [
 	  "eslint:recommended",
 	  "prettier"
   ],
+  "plugins": ["jest"],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": 12,
