+++ .gitignore
@@ -107,3 +107,6 @@ dist
 
 # TernJS port file
 .tern-port
+
+# packr output
+output
